legacy_agent/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pytest
aiohttp
anthropic
rich
tavily-python
ollama
groq
openai
google-generativeai
psycopg2-binary
llama-index
llama-index-vector-stores-postgres
pymongo
langchain
motor
sse-starlette
sh
gunicorn
jupyterlab
requests
crawl4ai @ git+https://github.com/unclecode/crawl4ai.git